# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HTMLPush API",
    "description": "HTMLPush 是一个让手机浏览器变成桌面显示器的服务。后台上传或切换 HTML 内容，手机端读取同一份 API 文档产物。",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "schemas": {
...
      "PageInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "页面唯一 ID"
          },
          "name": {
            "type": "string",
            "description": "页面名称"
          },
          "filename": {
            "type": "string",
            "description": "存储文件名"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "description": "上传时间"
          }
        },
        "required": [
          "id",
          "name",
          "filename",
          "uploadedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DeletedByIdResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "删除成功标记"
          },
          "id": {
            "type": "string",
            "description": "已删除资源 ID"
          }
        },
        "required": [
          "deleted",
          "id"
        ]
      },
      "CurrentPageResponse": {
        "type": "object",
        "properties": {
          "page": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PageInfo"
              },
              {
                "type": [
                  "object",
                  "null"
                ],
                "description": "当前展示页面"
              }
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ],
            "description": "当前页面 HTML 内容"
          }
        },
        "required": [
          "page",
          "content"
        ]
      },
      "SwitchCurrentPageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "切换成功标记"
          },
          "pageId": {
            "type": "string",
            "description": "当前页面 ID"
          },
          "timestamp": {
            "type": "integer",
            "description": "操作时间戳"
          }
        },
        "required": [
          "success",
          "pageId",
          "timestamp"
        ]
      },
      "SwitchCurrentPageRequest": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "string",
            "minLength": 1,
            "description": "要切换到的页面 ID"
          }
        },
        "required": [
          "pageId"
        ]
      },
...
      "ImageAsset": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ImageAssetBase"
          },
          {
            "type": "object",
            "properties": {
              "imageUrl": {
                "type": "string",
                "description": "原图访问地址"
              },
              "pageUrl": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "关联页面访问地址"
              }
            },
            "required": [
              "imageUrl",
              "pageUrl"
            ]
          }
        ]
      },
      "ImageAssetBase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "图片唯一 ID"
          },
          "name": {
            "type": "string",
            "description": "图片名称"
          },
          "filename": {
            "type": "string",
            "description": "存储文件名"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME 类型"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "文件大小，单位字节"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "description": "上传时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          },
          "pageId": {
            "type": [
              "string",
              "null"
            ],
            "description": "关联页面 ID"
          }
        },
        "required": [
          "id",
          "name",
          "filename",
          "mimeType",
          "size",
          "uploadedAt",
          "updatedAt",
          "pageId"
        ]
      },
...
      "ImageRenameRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "新的图片名称"
          }
        },
        "required": [
          "name"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          }
        },
        "required": [
          "success"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 1,
            "description": "管理密码"
          }
        },
        "required": [
          "password"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/current": {
      "get": {
        "tags": [
          "展示控制"
        ],
        "summary": "获取当前展示的页面信息",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentPageResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "展示控制"
        ],
        "summary": "切换当前展示的页面",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchCurrentPageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "切换成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwitchCurrentPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "缺少 pageId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "页面不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "切换失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/{id}": {
      "get": {
        "tags": [
          "图床管理"
        ],
        "summary": "获取原始图片内容",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "资源 ID",
              "example": "1776050757560-2g19q1"
            },
            "required": true,
            "description": "资源 ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "图片不存在"
          }
        }
      },
      "patch": {
        "tags": [
          "图床管理"
        ],
        "summary": "重命名图片",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "资源 ID",
              "example": "1776050757560-2g19q1"
            },
            "required": true,
            "description": "资源 ID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageAsset"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "图片不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "重命名失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "图床管理"
        ],
        "summary": "删除单张图片",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "资源 ID",
              "example": "1776050757560-2g19q1"
            },
            "required": true,
            "description": "资源 ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "图片不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "删除失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "管理后台登录",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功，设置 cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "请输入密码",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "登录失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}